tests/wpt/web-platform-tests/XMLHttpRequest/no-utf16-json.htm

<!doctype html>
<meta charset=utf-8>
<title>XMLHttpRequest: responseType "json" should be decoded as UTF-8 even when there's a UTF-16 BOM</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://infra.spec.whatwg.org/#parse-json-from-bytes" data-tested-assertations="following::ol/li[1]" />
<div id="log"></div>

<script>
  async_test(function() {
    var client = new XMLHttpRequest();
    client.responseType = 'json';
    client.onload = this.step_func_done(function(e) {
      assert_equals(client.response, null);
    });
    client.open("GET", "resources/utf16-bom.json");
    client.send(null);
  }, 'JSON asa UTF-16 with BOM should decode as UTF-8 (and, therefore, fail)');
</script>
